Add /opt/vde/bin/vde_vmnet and prerequites to lima-and-qemu tarball
It is kind-of heavy-handed by hardcoding the 2 dependencies, but
extending the opensnoop processing just to deal with /opt/vde would
be rather convoluted (it would only cover the dylib; the executables
have to be added manually anyways).

Installation will have to move the `vde` directory from the tarball
into a location only root can write to.

+++ hack/lima-and-qemu.pl
@@ -115,9 +115,28 @@ for my $file (keys %resign) {
     system("codesign --sign - --force --preserve-metadata=entitlements $file");
 }
 
-unlink("$repo_root/$dist.tar.gz");
 my $files = join(" ", map s|^$install_dir/||r, keys %deps);
+
+# Package vde_vmnet and prerequisites from /opt/vde → vde in the tarball
+my $opt_vde = "/opt/vde";
+die if -e "/tmp/$dist/vde";
+if (-f "$opt_vde/bin/vde_vmnet") {
+    my $dylib = "libvdeplug.3.dylib";
+    system("mkdir -p /tmp/$dist/vde/lib");
+    system("cp $opt_vde/lib/$dylib /tmp/$dist/vde/lib/$dylib");
+    $files .= " vde/lib/$dylib";
+
+    system("mkdir -p /tmp/$dist/vde/bin");
+    for my $tool (qw(vde_switch vde_vmnet)) {
+        system("cp $opt_vde/bin/$tool /tmp/$dist/vde/bin/$tool");
+        system "install_name_tool -change $opt_vde/lib/$dylib \@executable_path/../$dylib /tmp/$dist/vde/bin/$tool 2>&1";
+        $files .= " vde/bin/$tool";
+    }
+}
+
+unlink("$repo_root/$dist.tar.gz");
 system("tar cvfz $repo_root/$dist.tar.gz -C /tmp/$dist $files");
+
 exit;
 
 # File references may involve multiple symlinks that need to be recorded as well, e.g.
